一键划转 token币转出地址错误bug修复 btc转账失败bug修复

+++ app/src/main/java/com/cdkj/token/model/RedPacketDetails.java
@@ -117,7 +117,7 @@ public class RedPacketDetails {
         this.greeting = greeting;
     }
 
-    public double getSendNum() {
+    public int getSendNum() {
         return sendNum;
     }
